+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Reject blank activation name in activation rename [(#739)](https://github.com/wultra/powerauth-restful-integration/issues/739)
+
 ## [2.2.0] - 2026-07-20
 
 ### Added